// src/app/@core/models/site-inspection.ts
export class SiteInspection {
    date: any;
    ref_number: any;
    approved_by: any;
    site: any;
    contractor: any;
    site_engineer: any;
    work_description: any;
    ppe: any;
    risk_treatment: any;
    compliance_to_safaricom: any;
    provision_of_work_equipment: any;
    emergency_response: any;
}